revert non-conforming changes to public interfaces
- rotation for pattern
these changes were straight against the spec. next commit will reintroduce same
feature in a compatible manner (with transform-region).

- get rid of custom medium-draw-pattern
We are going to implement it properly with inks without custom methods which
only complicate the design. This commit breaks image renderer temporarily
because the renderer depends on draw-pattern* which draws with a non-uniform
ink (image), next commit will implement the missing functionality.

- medium-draw-text* doesn't take non-conforming argument TRANSFORMATION

We do not pass transformation as an argument to medium-draw-text* against the
specification. We record it though and take it from medium-translation which is
bound in reply-fn.

- introduce correct implementation of draw-pattern* and draw-design pattern

This change assumes, that backend is able to conformingly transform inks. We add
transformed-pattern class to be able to draw transformed-patterns.

- cosmetic: generate-medium-recording: prevent empty let in expansion this only
modifies how macro expands.